feat: Add GetOwnerAsync method to Character, Company, Inventory, and Sanction classes
This commit introduces a new feature by adding the GetOwnerAsync method to the Character, Company, Inventory, and Sanction classes. The GetOwnerAsync method retrieves the owner of each respective class asynchronously using the OwnerId property. This change enhances the functionality of these classes by providing an easy way to retrieve the owner information.